+++ Sail/src/Network/NetworkModule.hpp
@@ -6,12 +6,7 @@
 //#include <iostream>
 #include <thread>
 #include <mutex>
-
-#define MAX_PACKAGE_SIZE 64
-#define MAX_AWAITING_PACKAGES 1000
-//#define DEBUG_NETWORK
-
-typedef unsigned long long ConnectionID;
+#include "NetworkStructs.hpp"
 
 struct Connection
 {
@@ -26,30 +21,6 @@ struct Connection
 		socket = NULL;
 		ip = port = "";
 		thread = nullptr;
-	}
-};
-
-enum class NETWORK_EVENT_TYPE {
-	NETWORK_ERROR,
-	CLIENT_JOINED,
-	CLIENT_DISCONNECTED,
-	CLIENT_RECONNECTED,
-	MSG_RECEIVED,
-};
-
-struct MessageData {
-	char msg[MAX_PACKAGE_SIZE];
-};
-
-struct NetworkEvent {
-	NETWORK_EVENT_TYPE eventType;
-	ConnectionID clientID;
-	MessageData* data;
-
-	NetworkEvent() {
-		eventType = NETWORK_EVENT_TYPE::NETWORK_ERROR;
-		clientID = 0;
-		data = nullptr;
 	}
 };
